refactor: 검색 창 위치 조절
