+++ src/components/recruitment/recruitment_map/RecruitmentFilter.module.scss
@@ -1,9 +1,17 @@
 .filterContainer {
   display: flex;
-  align-items: center;
-  padding: 0.75rem 1.5rem;
-  background-color: #f8f9fa;
-  gap: 0.75rem;
+  gap: 0.5rem;
+  padding: 1rem;
+  background-color: #fff;
+  border-bottom: 1px solid #e9ecef;
+  overflow-x: auto;
+  white-space: nowrap;
+
+  &::-webkit-scrollbar {
+    display: none; /* Chrome, Safari, Opera*/
+  }
+  -ms-overflow-style: none; /* IE and Edge */
+  scrollbar-width: none; /* Firefox */
 }
 
 .filterGroup {
