+++ src/components/recruitment/recruitment_map/RecruitmentSearch.module.scss
@@ -1,0 +1,57 @@
+.searchContainer {
+  display: flex;
+  justify-content: center;
+  padding: 1rem 1rem;
+  background-color: #f8f9fa;
+}
+
+.searchWrapper {
+  position: relative;
+  width: 100%;
+  max-width: 600px;
+}
+
+.searchInput {
+  width: 100%;
+  padding: 0.5rem 1.5rem 1rem 1.5rem;
+  font-size: 1rem;
+  border: 1px solid #dee2e6;
+  border-radius: 50px;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.05);
+  transition:
+    border-color 0.2s,
+    box-shadow 0.2s;
+
+  &:focus {
+    outline: none;
+    border-color: #4c6ef5;
+    box-shadow: 0 0 0 3px rgba(76, 110, 245, 0.2);
+  }
+
+  &::placeholder {
+    color: #adb5bd;
+  }
+}
+
+.searchButton {
+  position: absolute;
+  right: 8px;
+  top: 50%;
+  transform: translateY(-50%);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 40px;
+  height: 40px;
+  background-color: #4c6ef5;
+  color: white;
+  border: none;
+  border-radius: 50%;
+  cursor: pointer;
+  font-size: 1rem;
+  transition: background-color 0.2s;
+
+  &:hover {
+    background-color: #3b5bdb;
+  }
+}
